[editor] Rename autoBox to boxer
