+++ packages/editor/src/connects/graph.js
@@ -1,13 +1,13 @@
 import { connect } from 'react-redux';
 import { updateBoxes } from '../actions';
 
-const mapDispatchToProps = autoBox => dispatch => {
+const mapDispatchToProps = boxer => dispatch => {
   const handlers = {
     onMouseDown: event => dispatch({ type: 'mouseDown', event }),
     onMouseMove: event => dispatch({ type: 'mouseMove', event }),
     onMouseUp: event => dispatch({ type: 'mouseUp', event }),
   };
-  if (autoBox) {
+  if (boxer) {
     handlers.onBoxes = boxes => dispatch(updateBoxes(boxes));
   }
   return handlers;
@@ -20,8 +20,8 @@ const mapStateToProps = ({ nodes, boxes, connections, boxRequests }) => ({
   ...(boxRequests.length && { boxRequests }),
 });
 
-export default ({ autoBox } = {}) =>
+export default ({ boxer } = {}) =>
   connect(
     mapStateToProps,
-    mapDispatchToProps(autoBox)
+    mapDispatchToProps(boxer)
   );
