+++ packages/graph/src/highs/index.js
@@ -1,6 +1,6 @@
 export * from './BoxContext';
 
-export { default as useAutoBox } from './useAutoBox';
+export { default as useBoxer } from './useBoxer';
 export { default as useAutoConnectionId } from './useAutoConnectionId';
 export { default as useAutoViewportSize } from './useAutoViewportSize';
 export { default as useBoxExtractor } from './useBoxExtractor';
@@ -11,7 +11,7 @@ export { default as useLoomer } from './useLoomer';
 export { default as useNodeDefaults } from './useNodeDefaults';
 export { default as useNormaliseConnections } from './useNormaliseConnections';
 
-export { default as withAutoBox } from './withAutoBox';
+export { default as withBoxer } from './withBoxer';
 export { default as withInteraction } from './interaction/withInteraction';
 export { default as withSvgPosition } from './withSvgPosition';
 export { default as withHtmlPosition } from './withHtmlPosition';
